wild-fire-mape2/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.1.0
redis==5.0.1
scikit-learn==1.3.2
geopy==2.4.0
python-geohash==0.8.5
pydantic==2.5.0
pydantic-settings==2.1.0
aiohttp==3.9.1
celery==5.3.4
python-dotenv==1.0.0